requirements.txt
aiogram
asyncio
aiohttp
requests
pymongo
python-dotenv
colorama